requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.2.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
narwhals==2.2.0
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pandera==0.26.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
# pywin32==311
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
narwhals==2.2.0
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pandera==0.26.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
python-calamine==0.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
narwhals==2.2.0
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pandera==0.26.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
python-calamine==0.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
